feat: add description for VIP 328
