+++ multisig/proposals/ethereum/vip-038/index.ts
@@ -6,7 +6,7 @@ export const COMPTROLLER = "0x687a01ecF6d3907658f7A7c714749fAC32336D1B";
 export const vsFRAX = "0x17142a05fe678e9584FA1d88EfAC1bF181bF7ABe";
 export const BORROW_CAP = parseUnits("2000000", 18);
 
-export const vip037 = () => {
+export const vip038 = () => {
   return makeProposal([
     {
       target: COMPTROLLER,
@@ -16,4 +16,4 @@ export const vip037 = () => {
   ]);
 };
 
-export default vip037;
+export default vip038;
